+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.29.0] - 2021-08-19
+
+- Added global icon
+
 ## [1.28.1] - 2021-08-19
 
 - Added year to DatePicker month label
