+++ src/components/Icons/Information/index.tsx
@@ -1,11 +1,18 @@
 import * as React from 'react';
 
-const Information: React.FC<React.SVGProps<SVGSVGElement>> = ({ ...props }) => (
-  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill='none' {...props}>
-    <circle cx="12" cy="12" r="9" stroke="currentColor" strokeWidth="2" fill="none" />
-    <rect x="13" y="17" width="2" height="7" rx="1" transform="rotate(-180 13 17)" fill="currentColor" stroke='none' />
-    <rect x="13" y="9" width="2" height="2" rx="1" transform="rotate(-180 13 9)" fill="currentColor" stroke='none' />
+const Global: React.FC<React.SVGProps<SVGSVGElement>> = ({ ...props }) => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="32" height="32" viewBox="0 0 32 32" fill="none" {...props}>
+    <path
+      d="M13.3115 23.3334H5.86573C4.92488 23.3334 4.15954 22.5679 4.15954 21.6272V8.05758C4.15954 7.15859 4.89108 6.42696 5.79015 6.42696H9.24552L12.0993 9.15823C12.2383 9.29144 12.4232 9.36553 12.6154 9.36553H24.1126C25.0116 9.36553 25.7432 10.0971 25.7432 10.9961C25.7432 11.4085 26.0773 11.7426 26.4897 11.7426C26.902 11.7426 27.2361 11.4085 27.2361 10.9961C27.2361 9.27393 25.8348 7.87263 24.1126 7.87263H12.9149L10.0612 5.14136C9.9222 5.00824 9.73734 4.93414 9.54512 4.93414H5.79015C4.06793 4.93414 2.66669 6.33536 2.66669 8.05758V21.6272C2.66669 23.3912 4.10165 24.8262 5.86573 24.8262H13.3115C13.7239 24.8262 14.058 24.4921 14.058 24.0797C14.058 23.6675 13.7239 23.3334 13.3115 23.3334Z"
+      fill="#100F0D"
+    />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M14.6667 19.7376C14.6667 15.6896 17.9447 12.4043 21.9927 12.4043C26.048 12.4043 29.3334 15.6896 29.3334 19.7376C29.3334 23.7856 26.048 27.071 21.9927 27.071C17.9447 27.071 14.6667 23.7856 14.6667 19.7376ZM24.9114 16.8043H27.0747C26.3707 15.5943 25.2487 14.6556 23.8994 14.1936C24.3394 15.0076 24.6767 15.8876 24.9114 16.8043ZM21.9998 13.9003C22.6085 14.7803 23.0852 15.7557 23.4005 16.8043H20.5992C20.9145 15.7557 21.3912 14.7803 21.9998 13.9003ZM16.1332 19.7376C16.1332 20.2436 16.2066 20.7349 16.3239 21.2043H18.8026C18.7439 20.7203 18.6999 20.2363 18.6999 19.7376C18.6999 19.2389 18.7439 18.7549 18.8026 18.2709H16.3239C16.2066 18.7403 16.1332 19.2316 16.1332 19.7376ZM16.9255 22.6709H19.0888C19.3235 23.5876 19.6608 24.4676 20.1008 25.2816C18.7515 24.8196 17.6295 23.8883 16.9255 22.6709ZM16.9255 16.8043H19.0888C19.3235 15.8876 19.6608 15.0076 20.1008 14.1936C18.7515 14.6556 17.6295 15.5869 16.9255 16.8043ZM21.9998 25.5749C21.3912 24.6949 20.9145 23.7196 20.5992 22.6709H23.4005C23.0852 23.7196 22.6085 24.6949 21.9998 25.5749ZM20.284 21.2043H23.716C23.782 20.7203 23.8334 20.2363 23.8334 19.7376C23.8334 19.2389 23.782 18.7476 23.716 18.2709H20.284C20.218 18.7476 20.1667 19.2389 20.1667 19.7376C20.1667 20.2363 20.218 20.7203 20.284 21.2043ZM23.8994 25.2816C24.3394 24.4676 24.6767 23.5876 24.9114 22.6709H27.0747C26.3707 23.8809 25.2487 24.8196 23.8994 25.2816ZM25.2995 19.7376C25.2995 20.2363 25.2555 20.7203 25.1968 21.2043H27.6755C27.7928 20.7349 27.8662 20.2436 27.8662 19.7376C27.8662 19.2316 27.7928 18.7403 27.6755 18.2709H25.1968C25.2555 18.7549 25.2995 19.2389 25.2995 19.7376Z"
+      fill="currentColor"
+    />
   </svg>
 );
 
-export default Information;
+export default Global;
