+++ src/components/Icons/index.ts
@@ -32,3 +32,4 @@ export { default as Chain } from './Chain';
 export { default as Information } from './Information';
 export { default as Calendar } from './Calendar';
 export { default as Filter } from './Filter';
+export { default as Global } from './Global';
